refactor(host)!: Rename Agent API namespace to board
- Move public SDK headers from librmcs/agent to librmcs/board
- Rename librmcs::agent to librmcs::board across host code
- Update internal includes to use the new board API

BREAKING CHANGE: Public host SDK includes and namespace references now use board instead of agent.
